# src/main/java/com/guibaarros/fiap/postech/fastfood/infrastructure/web/exceptionhandler/ErrorDTO.java
package com.guibaarros.fiap.postech.fastfood.infrastructure.web.exceptionhandler;

import com.fasterxml.jackson.databind.PropertyNamingStrategies;
import com.fasterxml.jackson.databind.annotation.JsonNaming;
import lombok.AllArgsConstructor;
import lombok.Getter;
import lombok.NoArgsConstructor;

@NoArgsConstructor
@AllArgsConstructor
@JsonNaming(PropertyNamingStrategies.SnakeCaseStrategy.class)
@Getter
public class ErrorDTO {
    private String errorMessage;
}